requirements.txt
streamlit
openai
requests
json
urllib
pillow
matplotlib